Organize benchmark tooling and results
