+++ pyproject.toml
@@ -38,11 +38,11 @@ main.cmd = "python main.py"
 main-week1.cmd = "python main.py --loader week1"
 main-week2.cmd = "python main.py --loader week2"
 main-week3.cmd = "python main.py --loader week3"
-bench.cmd = "python bench.py"
-bench-week2-operators.cmd = "python bench_week2_operators.py"
-bench-course-progression.cmd = "python bench_course_progression.py"
-bench-week2-progression.cmd = "python bench_course_progression.py --suite week2"
-bench-serving-progression.cmd = "python bench_serving_progression.py"
+bench.cmd = "python -m benches.bench"
+bench-week2-operators.cmd = "python -m benches.bench_week2_operators"
+bench-course-progression.cmd = "python -m benches.bench_course_progression"
+bench-week2-progression.cmd = "python -m benches.bench_course_progression --suite week2"
+bench-serving-progression.cmd = "python -m benches.bench_serving_progression"
 batch-main.cmd = "python batch-main.py"
 agent.cmd = "python agent.py"
 test.cmd = "python scripts/dev-tools.py test"
